requirements.txt
# C盘清理工具依赖

# 核心依赖
pydantic>=2.6.3
loguru>=0.7.2
psutil>=5.9.0
apscheduler>=3.10.0

# 数据库
pyyaml>=6.0

# AI / LLM
google-generativeai>=0.5.0

# 图片处理
imagehash>=4.3.1
Pillow>=10.0.0

# GUI (可选，用于未来GUI版本)
# PyQt6>=6.4.0
# PySide6>=6.4.0

# 添加报告功能所需的依赖
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.1
matplotlib==3.8.3
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
bcrypt==4.1.2
typing-extensions==4.10.0